+++ CHANGELOG.md
@@ -1,6 +1,24 @@
 # Changelog
 All Notable changes to `jobs-dice` will be documented in this file
 
+## 0.3.0 - 2015-09-28
+
+### Added
+- Support for all setter methods outlined in the [Dice API](http://www.dice.com/common/content/util/apidoc/jobsearch.html)
+- Readme documentation for all supported methods
+
+### Deprecated
+- Nothing
+
+### Fixed
+- Nothing
+
+### Removed
+- Nothing
+
+### Security
+- Nothing
+
 ## 0.2.4 - 2015-08-12
 
 ### Added